requirements.txt
absl-py==2.2.2
acres==0.5.0
aioice==0.9.0
aiortc==1.11.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
av==14.2.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chromadb==1.0.12
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.2.0
cryptography==44.0.2
distro==1.9.0
dnspython==2.7.0
durationpy==0.10
etelemetry==0.3.1
exceptiongroup==1.3.0
fastapi==0.115.9
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fpdf==1.7.2
fsspec==2025.5.1
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-crc32c==1.7.1
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==2.15.0
kubernetes==33.1.0
libclang==18.1.1
looseversion==1.3.0
lxml==5.4.0
lz4==4.4.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.2.0
mmh3==5.1.0
mpmath==1.3.0
mtcnn==1.0.0
namex==0.0.8
narwhals==1.33.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.95.1
opencv-python==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pdfrw==0.4
pillow==11.1.0
posthog==4.8.0
praw==7.8.1
prawcore==2.4.0
proto-plus==1.26.1
protobuf==5.29.5
prov==2.0.2
puremagic==1.29
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pydeck==0.9.1
pydot==4.0.0
pyee==13.0.0
Pygments==2.19.1
pylibsrtp==0.11.0
PyMuPDF==1.26.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
streamlit==1.44.1
streamlit-webrtc==0.56.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_keras==0.1
termcolor==3.0.1
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.7.1
tornado==6.4.2
tqdm==4.67.1
traits==7.0.2
transformers==4.52.4
typer==0.16.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.3
watchdog==6.0.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.14.1
zipp==3.23.0
